adjust pil size
