+++ client/src/components/OrgSettings.tsx
@@ -6,7 +6,6 @@ import { NotFound } from "@/components/NotFound";
 import { UpdateOrganizationDialog } from "@/components/OrganizationDialog";
 import { ProjectsTable } from "@/components/projects/ProjectsTable";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { DialogFooter } from "@/components/ui/dialog";
 import {
@@ -20,23 +19,16 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { UserTable } from "@/components/users/UserTable";
-import { Tier, UserRole } from "@/types/types";
+import { UserRole } from "@/types/types";
 import { useDeleteOrganizationMutation } from "@/utils/organizations";
 import { userQueryOptions } from "@/utils/users";
-import { isLilypadCloud } from "@/ee/utils/common";
-import { licenseQueryOptions } from "@/ee/utils/organizations";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
 import { Trash, TriangleAlert } from "lucide-react";
 import { Dispatch, SetStateAction } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
-
-const tier = {
-  [Tier.FREE]: "Free",
-  [Tier.PRO]: "Pro",
-  [Tier.TEAM]: "Team",
-};
+import { PlanList } from "@/components/PlanList.tsx";
 
 interface OrgSettingsProps {
   open: boolean;
@@ -44,67 +36,13 @@ interface OrgSettingsProps {
 }
 export const OrgSettings = ({ open, setOpen }: OrgSettingsProps) => {
   const { data: user } = useSuspenseQuery(userQueryOptions());
-  const { data: licenseInfo } = useSuspenseQuery(licenseQueryOptions());
   const activeUserOrg = user.user_organizations?.find(
     (userOrg) => userOrg.organization_uuid === user.active_organization_uuid
   );
   if (!activeUserOrg) return <NotFound />;
   return (
     <>
-      <div className="mb-6">
-        <div className="mb-4">
-          <span className="text-sm font-medium block mb-2">Available Plans:</span>
-          <div className="flex flex-wrap gap-2">
-            {Object.entries(tier).map(([tierKey, tierName]) => {
-              const tierNumber = Number(tierKey);
-              const isCurrentPlan = licenseInfo.tier.valueOf() === tierNumber;
-              const deploymentType = isLilypadCloud() ? "Cloud" : "Self-Host";
-              const fullPlanName = `${deploymentType} ${tierName}`;
-              const isTeamOrProPlus = tierNumber === Tier.TEAM.valueOf();
-
-              return (
-                <div key={tierKey} className="inline-block">
-                  {isCurrentPlan ? (
-                    <Badge pill variant="secondary" className="cursor-default">
-                      {fullPlanName} (Current)
-                    </Badge>
-                  ) : (
-                    <LilypadDialog
-                      title={`Upgrade to ${fullPlanName}`}
-                      description={
-                        isTeamOrProPlus
-                          ? `Contact us to upgrade your plan to ${tierName} during the beta period without billing.`
-                          : `Contact us to learn more about the ${tierName} plan.`
-                      }
-                      customTrigger={
-                        <Badge
-                          pill
-                          variant="outline"
-                          className="cursor-pointer hover:bg-gray-100"
-                        >
-                          {fullPlanName}
-                        </Badge>
-                      }
-                    >
-                      <div className="p-4">
-                        <p>
-                          Please email us at <a href="mailto:support@mirascope.com" className="text-blue-500 hover:underline">support@mirascope.com</a> to request an upgrade to the {tierName} plan.
-                        </p>
-                        {isTeamOrProPlus && (
-                           <p className="mt-2">
-                             During the beta period, this plan may be available without billing for eligible users.
-                           </p>
-                        )}
-                      </div>
-                    </LilypadDialog>
-                  )}
-                </div>
-              );
-            })}
-          </div>
-        </div>
-      </div>
-
+      <PlanList />
       <UpdateOrganizationDialog open={open} setOpen={setOpen} />
       <UserTable />
       <ProjectsTable />
